added HandleClick to server

+++ KvmSwitch.Shared/Models/MouseEvent.cs
@@ -35,8 +35,8 @@ namespace Shared
     {
         public uint ClickType { get; set; }
         public short ScrollSpeed { get; set; }
-        public float VelocityX { get; set; }
-        public float VelocityY { get; set; }
+        public int VelocityX { get; set; }
+        public int VelocityY { get; set; }
         public double TimeDelta { get; set; }
     }
 }
